split 'account.go' in 'user.go' and 'userinfo.go' in design
